Revert "[DF-5671] Used 64bits for components mask"
This reverts commit ae473b752d1de7cb33cf32b139f2950e0931c586.

+++ Sources/Internal/Entity/SceneSystem.cpp
@@ -47,7 +47,7 @@ SceneSystem::~SceneSystem()
     
 void SceneSystem::RegisterEntity(Entity * entity)
 {
-    uint64 requiredComponents = this->GetRequiredComponents();
+    uint32 requiredComponents = this->GetRequiredComponents();
     bool needAdd = ((requiredComponents & entity->GetAvailableComponentFlags()) == requiredComponents);
 
     if (needAdd)
@@ -56,7 +56,7 @@ void SceneSystem::RegisterEntity(Entity * entity)
     
 void SceneSystem::UnregisterEntity(Entity * entity)
 {
-    uint64 requiredComponents = this->GetRequiredComponents();
+    uint32 requiredComponents = this->GetRequiredComponents();
     bool needRemove = ((requiredComponents & entity->GetAvailableComponentFlags()) == requiredComponents);
 
     if (needRemove)
@@ -65,9 +65,9 @@ void SceneSystem::UnregisterEntity(Entity * entity)
     
 bool SceneSystem::IsEntityComponentFitsToSystem(Entity * entity, Component * component)
 {
-    uint64 entityComponentFlags = entity->GetAvailableComponentFlags();
-    uint64 componentToCheckType = 1 << component->GetType();
-    uint64 requiredBySystemComponents = this->GetRequiredComponents();
+    uint32 entityComponentFlags = entity->GetAvailableComponentFlags();
+    uint32 componentToCheckType = 1 << component->GetType();
+    uint32 requiredBySystemComponents = this->GetRequiredComponents();
 
     bool isAllRequiredComponentsAvailable = ((entityComponentFlags & requiredBySystemComponents) == requiredBySystemComponents);
     bool isComponentMarkedForCheckAvailable = ((requiredBySystemComponents & componentToCheckType) == componentToCheckType);
